DRTVWR-418: Fix a couple things that screw up 64-bit builds.
Do not replace 'x86' with 'i386' in specified arch: that produces the
nonsensical 'i386_64'. If you mean 'i386', say 'i386'. (We do.)

On Linux, do NOT force -m32 into the compile flags.

If we must force -Wall into the compile flags, put it at the start of the
list, not the end, so our passed-in CXXFLAGS can suppress individual warnings.
Otherwise we get spam without mercy.