DRTVWR-418: Replace buggy varargs functions with variadic templates.
Calls to makeStringArray() and makeStringList() blew up in 64-bit builds. Why?
Because they were passing things like:

makeStringList("1", "2", "3", "4", 0)

using va_arg(const char*) to walk the argument list. Why is that a problem?
Because in a 64-bit build, the int literal 0 is not as wide as const char*!

Because this repo is not properly set up for Vendor Branch strategy, I am
strenuously avoiding pervasive edits: edits that require patching every single
instance of a particular call. So changing every final '0' to
'static_cast<const char*>(0)' is out. Besides, new upstream calls wouldn't
follow that convention anyway.

Besides that, va_list and friends are crufty and type-unsafe and bad news and
should be eliminated whenever possible.

Instead, introduce makeStringArray() and makeStringList() implemented as
variadic templates, which is type-safe going forward. This requires a bit of
funny business to make them avoid treating the final '0' as a string.

This requires -std=c++11 in the compiler switches.